+++ CHANGELOG.md
@@ -7,6 +7,36 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Feature: Add new `LocationIcon` React component.
 - Feature: Add new `LocationInput` React component.
 
+## [7.0.0] - 2017-02-22
+
+Breaking changes:
+- The `FormAmountAndCurrency` has new props.
+  To be retro-compatible you should, in your app's components:
+  - Rename `value` to `currencyValue`.
+  - Rename `amount` to `currencyAmount`.
+  - Add the `tiny` boolean prop.
+  - Replace `import FormAmountAndCurrency` by
+    `import { FormAmountAndCurrency }`.
+- The `SelectWithState` has a new default size and export.
+  To be retro-compatible you should, in your app's components:
+  - Add the `tiny` boolean prop if the component does not have `big` set
+    to `true`.
+  - Remove the `big` boolean prop.
+
+Fixes:
+- Change font-size for `TextInputWithUnit` component.
+- Remove unknown props from `RadioButton`, `KarlHeader`,
+  `KarlHeaderWhenLogged`, `HeaderTitles` and `LoudspeakerIllustration`.
+
+## [6.4.0] - 2017-02-20
+
+Feature:
+- Add new `TextInputWithUnit` React component.
+
+Fixes:
+- Make unit from text input extensible according to its child.
+- Add missing props to `Tooltip` component.
+
 ## [6.3.0] - 2017-02-17
 
 Features: